Faster high_resolution_clock::now() for Windows
VS2017 ships with a high_resolution_clock::now() that works well, but takes
about 32 nanoseconds per call, because it's doing divide and mod. There's a
much faster way that involves precomputing the frequency as a 32.64 fixpoint
number. The resultant version runs in about 9 nanoseconds, which is a significant
improvement. TBD: write the precompute code (hard-coded constant for testing).