Bignum templated on word type
Need to do better tests, but it's possible that uint16_t and uint32_t
are roughly the same speed, despite uint32_t having 1/4th the
operations (because 64-bit multiply is slower than 32-bit multiply).

The real answer is for the multiplication primitive to be assembly,
so it can use 32x32=64 multiply and 64/32=32 divide.