Add exponentiation, right-shift, fix bug in multiply.
Fixed bug in multiply of the form n *= n (was returning zero due
to aliasing of lhs and rhs).

Added exponentiation. This is limited to N^(2^32-1)) for memory
reasons, and even that limit might be too high.

Added right-shift by arbitrary amount. Unlike ANSI C, right-shift
of a sufficiently large value results in zero, not undefined
behavior.