Rewrote buffer management code, more cleanup
The code is twice as big but has a more abstracted buffer class. I
still don't like it, because it's a little trickier to call than I
wanted, and I really want to reduce the amount of code so that it
stays readable while still being fast.

Also, tests need to be analytical instead of as cherry-picked
examples, so we can better prove the code is correct.